refactor(registration-react): Rename ConnectedFormDistribution to FormDistribution so that consumers can use canonical name
