+++ applications/registration-react/src/pages/dataset-registration-page/form-distribution/form-distribution.jsx
@@ -34,6 +34,6 @@ const mapStateToProps = (state, ownProps) => {
   };
 };
 
-export const ConnectedFormDistribution = connect(mapStateToProps)(
+export const FormDistribution = connect(mapStateToProps)(
   ConfiguredFormDistribution
 );
