+++ applications/registration-react/src/pages/dataset-registration-page/form-distribution/form-distribution.test.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { createMockStore } from 'redux-test-utils';
 import shallowWithStore from '../../../../test/shallowWithStore';
-import { ConnectedFormDistribution } from './connected-form-distribution.component';
+import { FormDistribution } from './form-distribution';
 import dataset from '../../../../test/fixtures/datasets';
 import openlicenses from '../../../../test/fixtures/openlicenses';
 
-test('should render ConnectedFormDistribution correctly', () => {
+test('should render FormDistribution correctly', () => {
   const testState = {};
   const store = createMockStore(testState);
   const datasetItem = dataset.datasetItems._embedded.datasets[0];
   const { openLicenseItems } = openlicenses;
 
   const wrapper = shallowWithStore(
-    <ConnectedFormDistribution
+    <FormDistribution
       datasetItem={datasetItem}
       openLicenseItems={openLicenseItems}
     />,
